Update openshift-ansible to release-1.3 branch
It looks like the openshift-ansible branch enterprise-XX
no longer exists.  For that reason, we have changed to the
release-1.3 branch.  But instead of just using the branch
name, we are using the commit hash to *try* to guarantee
that it will work for all.

Also updated to use EPEL release of Heketi instead of the
one from COPR.  Some users have complained that the
COPR repo times out.